Bugfix to show credits properly & select least loaded teacher. Made some code optization and added Java docs Added queries to verifying the implementation

+++ scheduler-api/src/main/java/com/highschool/scheduler/dto/AcademicHistoryDTO.java
@@ -4,7 +4,7 @@ public record AcademicHistoryDTO(
         String semesterName,
         String courseName,
         String courseType,
-        int credits,
+        double credits,
         String status
 ) {
 }
